Add optional argument for quantity to get_note_list()
Adds an optional argument allowing the user to specify the exact
number of notes to return in the list. If the argument is not set
by the user then it is assumed to be "infinite", or "all" notes.

If the quantity is less than the NOTE_FETCH_LENGTH (currently
specified as 20) then a call is made for the exact number of notes
required.

Otherwise, the correct quantity is determined in the while loop that
looks at the "mark" key set by Simplenote."