Add modifydate to update_note() function
Changed notes don't seem to get a modified timestamp by the server by
default. This adds the 'modifydate' key to the note object when
updating a note, with the current time in milli-seconds since epoch as value.

---

Pull in this commit from simplenote.vim:

https://github.com/mrtazz/simplenote.vim/commit/f0fa5c80c35d470779a76ffa728d1fc9f510c5dc

Note, that I'm (@atomicules) not convinced this is actually needed as per
the API docs, but working towards getting simplenote.py and
simplenote.vim aligned again so makes sense to include it here for now.

Needed to add `import time`