Rework the get_list_length_25 test
As per commit dcf8b72b378cd5e8539c3f3bf4288ae3abd01e10 the original
intention of this test was for fetching a list of notes longer than the
NOTE_FETCH_LENGTH. This was originally set to 20, but changed in commit
ea48a79889decc79752d9e3c9833682e16851ba0 to 100, effectively making this
test redundant.

As it was, this test would fail sporadically anyway. First, it would
never actually create 25 notes, I suspect due to rate limiting, rather
ending up with around 8 in total. As to why the assertion would actually
fail I'm not exactly sure, but within the associated Simplenote account
there were some stuck notes that were not being cleared out or retrieved
properly via this test script. I manually cleared the account and now
this version of the test seems to work.

Since we are only interested in testing whether we can fetch a list of
notes greater than NOTE_FETCH_LENGTH it makes sense to monkey patch that
to a smaller value than 100 and ensure we have created more notes than
this and then simply assert whether the number of notes returned is
greater than the NOTE_FETCH_LENGTH.