Minor change to datetime import to suit inclusion as ext mod. in Simplenote.vim
Changes to suit upcoming modification to Simplenote.vim. Want to split
out the Simplenote.py code from Simplenote.vim so it can be included as
an external module. This will make it easier to maintain Simplenote.vim
and keep it up to date with any changes in Simplenote.py. It will also
allow unit testing to be developed for Simplenote.vim.

Whilst testing this, the datetime module import in Simplenote.py would
not work in the `from datetime import datetime` format when referencing
the Simplenote.py module from within Simplenote.vim. By using a straight
`import datetime` format it worked ok.

[ci skip]