Ensure update_note uses utf-8 encoding if it returns note content
I.e. bring it in line with what happens for get_note

I was seeing sporadic unicode errors in a command line tool I use that
uses simplenote.py and I *think* this occured with merged notes. So far
this commit seems to have solved the problems I was seeing. It also
passes all the simplenote.py tests.