Make tests work again with single simplenote instance
- Remove time and sleeps

I have added these in myself quite a few times in the past in an effort
to help the tests past, but I think it's purely coincidental when they
"fix" problems - the Simplenote API is more than capable of handling the
number and speed of the requests made during this test run.

- Move single simplenote instance creation to `setUpClass`

The issue identified in Issue #12 is a good one and quite correct: The
test were inadvertently creating lots of instances rather than re-using
one for the test run. However, I think if we are going to have a single
instance created it makes sense to do it in `setUpClass`, which gets run
once at the start, rather than conditionally in `setUp` which runs
before every test.

- Remove `try` wrapped around `setUp` ("General setup fail")

As it was masking the real cause of errors within `setUp`, but then...

- Remove `self.fail` from note creations in `setUp`

...I found the general setup failures were caused by note creations
failing. I like the idea of knowing the truth, that `setUp` partially
failed, but this is no reason to fail that whole test before we've even
got to the test (see further note below though) - that's why we've got
the conditional `self.initial_note_count` increments in for the actual
tests to check against. Otherwise we are never going to get a test run
to pass. Before backing all these out I even tried sleeps in-between the
note creations, but odd failures still ok.

This point perhaps highlights that improvements to the tests can be made
- are we sure that if note creation in `setUp` is sometimes failing that
the tests aren't passing as a result by mistake? I'd prefer to back sure
the actually tests are robust/accurate rather use `self.fail` in
`setUp`.

References issue #12