Temporary fix for Unicode test in Python 3
This way the test doesn't fail in Python 3. I don't yet know a proper
way to test if a string is utf-8 in Python 3 because aren't they all
anyway? So for now just checking the string is equal.

This lets me concentrate on adding back in utf-8 support under Python 2.

References issue #13