Remove expectedFailure so test suite passes under Python 3
As it says in the comments in the code. For some reason having an
unexpected success causes the whole test suite to return as FAILED even
though each individual test passes.

However, now we've gone to the single Simplenote instance it seems this
test passes pretty consistently anyway so we might not need the
expectedFailure.

At this point all tests are passing under 2 and 3, but will do some
further manual checks to make sure everything is ok before tidying
everything up and merging into master.

References issue #12
References issue #13