Be more consistent with usage of response and json.loads
As a result of trying to use the codecs library in place of the
`.decode('utf-8')` I realised the code wasn't structured particularly
consistently where `response = urllib2.urlopen(request)` and
`json.loads` are used.

I couldn't get `codecs.getreader` to work so I'm sticking with the
`.decode...`, but the consistency changes are worth keeping.

References issue #13