Bump to version 1.0.0
No features changes since 0.4.0, but this is stable and proven enough to
be v1 now. I was waiting until the switch to the Simperium API for
version 1, but they seem to be in no rush to move everyone over to that,
so instead make this version 1 against the old Simplenote API and then
version 2 can be against the Simperium API.

Also, since Simplenote.vim is also close to version 1, it would be nice
to have it use a version 1 simplenote.py.