Fix for auto HTML entity escaping introduced in API
It seems a [recent update][1] introduced a change where html entities
are always encoded (unless the note is flagged as markdown). Simplenote
handle the unescaping of this in their apps so we need to do the same
here. E.g. `<` will be returned as `&lt;`

It's pretty much down to luck how long any of this continues to work.
Pretty sure all their official apps use the Simperium API which they
never released and this one is from 2011.

References: https://github.com/mrtazz/simplenote.vim/issues/99

[1]: https://twitter.com/simplenoteapp/status/963588833051684864