Sort tags in get_note / update_note
Had this as a TODO in simplenote.vim:

https://github.com/mrtazz/simplenote.vim/blob/29d52cc5a15af2b3dd357ab5610148af0e350152/autoload/SimplenoteUtilities.py#L728

I.e. Needed to move the sorting to simplenote.py

I don't think there is a problem in sorting in update as well as get.
Added a test, but since the update is also getting sorted the test is
pretty much checking both at the same time.