Re-enable tag filtering. Tests pass in Py 2 and 3
Right back when I first started looking at Simperium stuff I was working
on the basis of the index only returning a version and note id. But we
can pull all info with the index (now, maybe before?) so we can filter
on tags again.

Can't _sensibly_ filter on modified date though. Poop.